+++ packages/browser-wallet-api-helpers/src/wallet-api-types.ts
@@ -118,6 +118,11 @@ interface MainWalletApi {
      * Note that any calls will throw an error, if the site is not connected with the wallet.
      */
     getGrpcClient(): ConcordiumGRPCClient;
+    /**
+     * Returns the genesis hash of the currently selected chain in the wallet.
+     * Returns undefined if the wallet is either locked or not set up by the user.
+     */
+    getSelectedChain(): Promise<string | undefined>;
 
     /**
      * Request that the user adds the specified tokens for a given contract to the wallet.