+++ packages/browser-wallet-message-hub/src/handlers.ts
@@ -174,6 +174,24 @@ export class ContentMessageHandler {
     }
 }
 
+type BroadcastOptions = {
+    /**
+     * Disable whitelist requirement for broadcast method
+     *
+     * @example
+     * handler.broadcast(EventType.ChainChanged, undefined, {requireWhitelist: false});
+     */
+    requireWhitelist?: boolean;
+    /**
+     * Callback to be run for each non-whitelisted tab. This runs even if whitelist is disabled.
+     *
+     * @example
+     * handler.broadcast(EventType.ChangeAccount, "1234", {nonWhitelistedTabCallback: (tab: chrome.tabs.Tab) => ...});
+     */
+    nonWhitelistedTabCallback?(tab: chrome.tabs.Tab): void;
+};
+const defaultBroadcastOptions: BroadcastOptions = { requireWhitelist: true };
+
 export class ExtensionsMessageHandler extends BaseMessageHandler<WalletMessage> {
     constructor(
         private connectedSites: { get(): Promise<Record<string, string[]> | undefined> },
@@ -200,21 +218,36 @@ export class ExtensionsMessageHandler extends BaseMessageHandler<WalletMessage> 
     /**
      * Broadcast an event of a specific type, with an optional payload, to all currently
      * open and whitelisted (connected to the selected account) tabs.
+     * By specifying options, it's possible to disable the whitelist by `{requireWhitelist: false}`
+     * and also declare a callback to be called for tabs not included in the whitelist through `{nonWhitelistedTabCallback: (t: chrome.tabs.Tab) => ...}`
+     * Default options are {requireWhitelist: true}
      *
      * @example
      * handler.broadcast(EventType.ChangeAccount, "1234");
+     * handler.broadcast(EventType.ChainChanged, undefined, {requireWhitelist: false});
+     * handler.broadcast(EventType.ChangeAccount, "1234", {nonWhitelistedTabCallback: handlerFunction});
      */
     // TODO would be nice to make this more type safe.
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    public broadcast(type: EventType, payload?: any): void {
+    public broadcast(type: EventType, payload?: any, options: BroadcastOptions = {}): void {
+        const optionsWithDefaults = { ...defaultBroadcastOptions, ...options };
         chrome.tabs
             .query({}) // get all
-            .then(this.getWhitelistedTabs.bind(this))
-            .then((tabs) =>
-                tabs
+            .then((ts) =>
+                this.getWhitelistedTabs(ts).then((wl) => ({
+                    valid: optionsWithDefaults.requireWhitelist ? wl : ts,
+                    invalid: ts.filter((t) => !wl.some((w) => w.id === t.id)),
+                }))
+            )
+            .then(({ valid, invalid }) => {
+                valid
                     .filter(({ id }) => id !== undefined)
-                    .forEach((t) => this.sendEventToTab(t.id as number, new WalletEvent(type, payload)))
-            );
+                    .forEach((t) => this.sendEventToTab(t.id as number, new WalletEvent(type, payload)));
+
+                if (optionsWithDefaults.nonWhitelistedTabCallback !== undefined) {
+                    invalid.forEach(optionsWithDefaults.nonWhitelistedTabCallback);
+                }
+            });
     }
 
     /**