+++ packages/browser-wallet/src/popup/store/account.ts
@@ -20,10 +20,10 @@ export const storedConnectedSitesAtom = atomWithChromeStorage<Record<string, str
 );
 
 const storedAccountAtom = atomWithChromeStorage<string | undefined>(ChromeStorageKey.SelectedAccount, undefined);
-export const selectedAccountAtom = atom<string | undefined, string | undefined>(
+export const selectedAccountAtom = atom<string | undefined, string | undefined, Promise<void>>(
     (get) => get(storedAccountAtom),
-    (_, set, address) => {
-        set(storedAccountAtom, address);
+    async (_, set, address) => {
+        await set(storedAccountAtom, address);
         popupMessageHandler.broadcast(EventType.AccountChanged, address);
     }
 );